Fix discussions, small refactor
