+++ lib/qonfig/plugins/vault/commands/definition/load_from_vault.rb
